moved training worker logic inside directory
