+++ src/bp/nlu-core/training-worker-queue/communication.ts
@@ -1,7 +1,7 @@
 import { NLU } from 'botpress/sdk'
 import { ErrorMessage } from 'ml/error-utils'
 
-import { TrainInput, TrainOutput } from './training-pipeline'
+import { TrainInput, TrainOutput } from '../training-pipeline'
 
 export type OutgoingPayload<T extends OutgoingMessageType> = T extends 'make_new_worker'
   ? { config: NLU.Config; requestId: string }
